+++ src/climate_data/special/__init__.py
@@ -2,6 +2,10 @@ from climate_data.special.compile_person_days import (
     compile_person_days,
     compile_person_days_task,
 )
+from climate_data.special.download_era5_uncertainty import (
+    download_era5_uncertainty,
+    download_era5_uncertainty_task,
+)
 from climate_data.special.temperature_person_days import (
     temperature_person_days,
     temperature_person_days_task,
@@ -15,10 +19,12 @@ RUNNERS = {
     "temperature_zone": generate_temperature_zone,
     "temperature_person_days": temperature_person_days,
     "compile_person_days": compile_person_days,
+    "download_era5_uncertainty": download_era5_uncertainty,
 }
 
 TASK_RUNNERS = {
     "temperature_zone": generate_temperature_zone_task,
     "temperature_person_days": temperature_person_days_task,
     "compile_person_days": compile_person_days_task,
+    "download_era5_uncertainty": download_era5_uncertainty_task,
 }